Use emaint for syncing and for applying package moves
By using emaint for syncing, EMERGE_DEFAULT_OPTS is ignored and, thus, it
will not break if a user has --ask as a default option (as some do). Note that
emerge --sync calls through to the emaint sync system anyway.

Also, emaint provides a specific option for applying package moves so use it.